move components to features
